+++ common/src/main/java/dev/mrsnowy/teleport_commands/suggestions/HomeSuggestionProvider.java
@@ -5,28 +5,36 @@ import com.mojang.brigadier.suggestion.SuggestionProvider;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.suggestion.SuggestionsBuilder;
 import dev.mrsnowy.teleport_commands.TeleportCommands;
-import dev.mrsnowy.teleport_commands.storage.StorageManager;
+
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
+
+import dev.mrsnowy.teleport_commands.storage.classes.NamedLocation;
+import dev.mrsnowy.teleport_commands.storage.classes.Player;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.server.level.ServerPlayer;
 
-import static dev.mrsnowy.teleport_commands.storage.StorageManager.GetPlayerStorage;
+import static dev.mrsnowy.teleport_commands.storage.StorageManager.STORAGE;
 
 public class HomeSuggestionProvider implements SuggestionProvider<CommandSourceStack> {
     @Override
     public CompletableFuture<Suggestions> getSuggestions(CommandContext<CommandSourceStack> context, SuggestionsBuilder builder) {
         try {
             ServerPlayer player = context.getSource().getPlayerOrException();
-            StorageManager.StorageClass.Player playerStorage = GetPlayerStorage(player.getStringUUID()).getSecond();
+            Optional<Player> optionalPlayerStorage = STORAGE.getPlayer(player.getStringUUID());
 
-            for (StorageManager.StorageClass.NamedLocation currentHome : playerStorage.Homes) {
-                builder.suggest(currentHome.name);
+            if (optionalPlayerStorage.isPresent()) {
+                Player PlayerStorage = optionalPlayerStorage.get();
+
+                for (NamedLocation currentHome : PlayerStorage.getHomes()) {
+                    builder.suggest(currentHome.getName());
+                }
             }
 
             // Build and return the suggestions
             return builder.buildFuture();
         } catch (Exception e) {
-            TeleportCommands.LOGGER.error("Error getting suggestions!");
+            TeleportCommands.LOGGER.error("Error getting home suggestions! ", e);
             return null;
         }
     }